refactor: fold app restoration into management agent
